requirements.txt
azure-storage-blob==12.26.0
azure-identity==1.25.0
uvicorn==0.37.0
fastapi==0.118.0